+++ app/policies/export_policy.rb
@@ -17,6 +17,13 @@ class ExportPolicy
     return true if @user.education_and_skills_funding_agency_team?
     return true if @user.academies_operational_practice_unit_team?
     return true if @user.business_support_team?
+    return true if @user.service_support_team?
+
+    false
+  end
+
+  def service_support?
+    return true if @user.service_support_team?
 
     false
   end