+++ spec/features/all_projects/export/export_users_can_see_an_export_landing_page_spec.rb
@@ -45,4 +45,16 @@ RSpec.feature "Export users can see the exports landing page" do
     expect(page).to have_content("academies due to transfer over the next 6 months")
     expect(page).to have_content("pre-transfer grants for academies joining a different trust")
   end
+
+  scenario "a service support user can see the exports landing page" do
+    user = create(:user, team: :service_support)
+
+    sign_in_with_user(user)
+    click_on "Exports"
+
+    expect(page).to have_content("funding agreement letter contacts, RPA and start-up grants")
+    expect(page).to have_content("pre-opening grants for schools becoming academies")
+    expect(page).to have_content("academies due to transfer over the next 6 months")
+    expect(page).to have_content("pre-transfer grants for academies joining a different trust")
+  end
 end